+++ mixinnet/rpc_utxo.go
@@ -9,7 +9,7 @@ type (
 		Type   uint8           `json:"type"`
 		Amount decimal.Decimal `json:"amount"`
 		Hash   Hash            `json:"hash"`
-		Index  uint64          `json:"index,omitempty"`
+		Index  uint8           `json:"index,omitempty"`
 		Lock   *Hash           `json:"lock,omitempty"`
 	}
 )
